+++ src/blocks/about.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { Content } from "../typeHelpers"
+import { Content } from "../libs/typeHelpers"
 import styled from "styled-components"
 
 type Props = {
